+++ app/about_ui/area51dialog.h
@@ -1,0 +1,12 @@
+#ifndef AREA51DIALOG_H
+#define AREA51DIALOG_H
+
+#include <QDialog>
+
+class Area51Dialog : public QDialog {
+  Q_OBJECT
+public:
+  explicit Area51Dialog(QWidget *parent = nullptr);
+};
+
+#endif // AREA51DIALOG_H